requirements.txt
PyYAML==3.10
nltk==3.0.2